fix: fw-1612 coin 24h price change
